requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-livereload-server==0.4
exceptiongroup==1.1.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
loguru==0.7.0
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
polib==1.2.0
prettyprint==0.1.5
pycparser==2.21
PyJWT==2.7.0
pytest==7.3.1
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.2
urllib3==2.0.2
whitenoise==6.5.0
